Refactor blackandwhitelist.py -> black_and_white_list.py
