+++ medusa/server/web/home/add_shows.py
@@ -18,7 +18,7 @@ from ..core import PageTemplate
 from .... import (
     classes, config, db, helpers, logger, ui,
 )
-from ....blackandwhitelist import short_group_names
+from ....black_and_white_list import short_group_names
 from ....common import Quality
 from ....helper.common import sanitize_filename, try_int
 from ....helpers import get_showname_from_indexer
